Correct a parameter for dielectric constatnt, MM_B.
There was a typo in the original Malmberg and Maryott 1956 (Eq. 5),
that was inherited in Hasted (1972). The correct value is -0.40008,
not -0.4008.

The deviation in the dielectric constant due to this error is zero at 0 °C
and appears slightly as the temperature rises:
0.018 (0.02%) at 25 °C, 0.036 (0.05%) at 50 °C, and 0.072 (0.13%) at 100 °C.
The effect of this difference should be negligible.